requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
open-clip-torch>=2.20.0

# Medical NLP
scispacy>=0.5.0
# Note: Run separately after install: pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_md-0.5.1.tar.gz

# Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Knowledge Graph
neo4j>=5.0.0
py2neo>=2021.1

# Vector Database
weaviate-client>=3.25.0

# LLM & Agents
langchain>=0.1.0
langgraph>=0.0.20
ollama>=0.1.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Monitoring
prometheus-client>=0.19.0

# Explainability
shap>=0.43.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.25.0
cryptography>=41.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0  # HuggingFace datasets for streaming
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Dashboard
streamlit>=1.28.0
